Added the settings link and registered activation deactivation hooks in the new organized plugin structure
