+++ public/wp-content/plugins/advThemeMang/advThemeMang.php
@@ -24,6 +24,17 @@ if(file_exists(dirname(__FILE__). '/vendor/autoload.php')){
 
 define('PLUGIN_PATH', plugin_dir_path(__FILE__));
 define('PLUGIN_URL',  plugin_dir_url(__FILE__));
+define('PLUGIN',  plugin_basename(__FILE__));
+
+function activate(){
+    Inc\Base\Activate::activate();
+}
+function deactivate() {
+    Inc\Base\Deactivate::deactivate();
+}
+
+register_activation_hook(__FILE__, 'activate');
+register_deactivation_hook(__FILE__, 'deactivate');
 
 if(class_exists('Inc\\Init')){
     Inc\Init::register_services();
