+++ public/wp-content/plugins/advThemeMang/inc/Base/Deactivate.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inc;
+namespace Inc\Base;
 class Deactivate {
     static function deactivate() {
         flush_rewrite_rules();
